navigate to projects page after creation
